+++ R/spinner.R
@@ -1,7 +1,5 @@
 #' spinner: a graph neural network leveraging both local and global representations
 #'
-#' spinner is ... .
-#'
 #' @docType package
 #' @name spinner
 "_PACKAGE"