+++ NEWS.md
@@ -1,3 +1,12 @@
 # spinner 1.0.0
 
 * Added a `NEWS.md` file to track changes to the package.
+
+# spinner 1.1.0
+
+* Message Passing Directionality: removed undirected option, name fix for the others
+* Managed the error case during automatic embedding of nodes/ edges
+* Automatic normalization (and de-normalization) for regression tasks
+* Fixed the issue caused by the name indexing in igraph (symbolic name indexes are removed at the very beginning of processing)
+* Added a specific function for hyper-parameter tuning based on random search
+